Working on changing dialog

+++ src/CDOption.hpp
@@ -8,13 +8,13 @@
 class CDOption
 { 
 private:
-    size_t m_ID;
+    int m_ID;
     std::string m_sText;
     std::string m_sTarget;
     bool m_active;
 
 public:
-    CDOption(size_t sID, std::string sText, std::string sTarget, bool active);
+    CDOption(int sID, std::string sText, std::string sTarget, bool active);
 
     // *** GETTER *** //
     std::string getText();
